+++ age-plugin/examples/age-plugin-unencrypted.rs
@@ -3,10 +3,11 @@ use age_core::{
     secrecy::ExposeSecret,
 };
 use age_plugin::{
+    Callbacks, PluginHandler,
     identity::{self, IdentityPluginV1},
     print_new_identity,
     recipient::{self, RecipientPluginV1},
-    run_state_machine, Callbacks, PluginHandler,
+    run_state_machine,
 };
 use clap::Parser;
 
@@ -21,7 +22,7 @@ const RECIPIENT_TAG: &str = PLUGIN_NAME;
 fn explode(location: &str) {
     if let Ok(s) = env::var("AGE_EXPLODES") {
         if s == location {
-            panic!("Env variable AGE_EXPLODES={} is set. Boom! 💥", location);
+            panic!("Env variable AGE_EXPLODES={location} is set. Boom! 💥");
         }
     }
 }